# docs/swagger.yaml
openapi: 3.0.0
info:
  title: PayKaduna Integration API
  version: 1.0.0
  description: |
    REST API gateway for PayKaduna Intelligent Billing System.
    
    Handles HMAC signing, request formatting, and error handling.
servers:
  - url: http://localhost:3000
    description: Local development server
  - url: https://api.example.com
    description: Production server
...
paths:
  /v1/bills:
    post:
      tags:
        - Bills
      summary: Create a single bill
      description: Create an electronic bill for a taxpayer
      operationId: createBill
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBillRequest'
            example:
              identifier: "TAXPAYER123"
              firstName: "John"
              middleName: "Doe"
              lastName: "Smith"
              telephone: "+2348012345678"
              address: "123 Main Street, Kaduna"
              engineCode: "185477"
              esBillDetailsDto:
                - amount: 10000.00
                  mdasId: 1
                  narration: "Property Tax"
                  recommendedRate: 0.05
      responses:
        '200':
          description: Bill created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateBillResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /v1/bills/bulk:
    post:
      tags:
        - Bills
      summary: Create bulk bills
      description: Create multiple bills in one request
      operationId: createBulkBills
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BulkBillRequest'
      responses:
        '200':
          description: All bills created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateBillResponse'
        '207':
          description: Partial success, some bills failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateBillResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /v1/bills/{reference}:
    get:
      tags:
        - Bills
      summary: Get bill by reference
      description: Get bill details by reference number
      operationId: getBill
      parameters:
        - name: reference
          in: path
          required: true
          schema:
            type: string
          description: Bill reference number
      responses:
        '200':
          description: Bill retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Bill'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /v1/bills/{reference}/invoice-url:
    get:
      tags:
        - Bills
      summary: Get invoice URL
      description: Get payment invoice URL for a bill
      operationId: getInvoiceUrl
      parameters:
        - name: reference
          in: path
          required: true
          schema:
            type: string
          description: Bill reference number
      responses:
        '200':
          description: Invoice URL retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvoiceUrlResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /v1/bills/{reference}/metadata:
    post:
      tags:
        - Bills
      summary: Attach metadata to bill
      description: Add custom metadata to a bill
      operationId: attachMetadata
      parameters:
        - name: reference
          in: path
          required: true
          schema:
            type: string
          description: Bill reference number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AttachDataRequest'
      responses:
        '200':
          description: Metadata attached successfully
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /v1/bills/metadata/bulk:
    post:
      tags:
        - Bills
      summary: Bulk attach metadata
      description: Add custom metadata to multiple bills
      operationId: bulkAttachMetadata
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BulkAttachDataRequest'
      responses:
        '200':
          description: Metadata attached
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /api/v1/paykaduna/webhook:
    post:
      tags:
        - Webhooks
      summary: Webhook endpoint
      description: |
        Receive webhook events from PayKaduna.
        
        **Authentication:**
        Include HMAC-SHA512 signature in `x-paykaduna-signature` header.
        
        **Signature:**
        Generate using: `HMAC-SHA512(JSON.stringify(requestBody), secretKey).hex()`
        Use the exact JSON string from the request body.
        
        **Request Format:**
        Include `event`, `data` (with `invoiceNo` as string), and `message` fields.
        
        **Event Types:**
        - `charge.success` or `payment.success`: Payment completed
        - `payment.error`: Payment failed
        - Custom: Any string value
      operationId: handleWebhook
      parameters:
        - name: x-paykaduna-signature
          in: header
          required: true
          schema:
            type: string
          description: HMAC-SHA512 signature
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WebhookEvent'
            examples:
              paymentSuccess:
                summary: Payment success event
                value:
                  event: "charge.success"
                  data:
                    invoiceNo: "INV123456"
                    billReference: "BILL123456"
                    amount: 10000.00
                    status: "paid"
                    transactionId: "TXN789"
                    timestamp: "2025-01-15T10:30:00Z"
                  message: "Payment completed successfully"
              paymentError:
                summary: Payment error event
                value:
                  event: "payment.error"
                  data:
                    invoiceNo: "INV123456"
                    billReference: "BILL123456"
                    transactionId: "TXN789"
                  message: "Payment processing failed"
      responses:
        '200':
          description: Webhook processed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookEvent'
              example:
                event: "charge.success"
                data:
                  invoiceNo: "INV123456"
                  billReference: "BILL123456"
                  amount: 10000.00
                  status: "paid"
                message: "Webhook event processed successfully"
        '400':
          description: Invalid request format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookEvent'
              example:
                event: "webhook.error"
                data: {}
                message: "Invalid request: event is required and must be a string"
        '401':
          description: Invalid or missing signature
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookEvent'
              examples:
                missingSignature:
                  value:
                    event: "webhook.error"
                    data: {}
                    message: "Webhook signature is required"
                invalidSignature:
                  value:
                    event: "webhook.error"
                    data: {}
                    message: "Webhook signature validation failed"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookEvent'
              example:
                event: "webhook.error"
                data: {}
                message: "Internal server error"
...
components:
  schemas:
    CreateBillRequest:
      type: object
      required:
        - identifier
        - firstName
        - lastName
        - telephone
        - address
        - esBillDetailsDto
      properties:
        identifier:
          type: string
          description: Taxpayer ID or TPUI
        firstName:
          type: string
        middleName:
          type: string
        lastName:
          type: string
        telephone:
          type: string
        address:
          type: string
        engineCode:
          type: string
          description: Optional engine code
        esBillDetailsDto:
          type: array
          items:
            $ref: '#/components/schemas/EsBillDetailDto'
    EsBillDetailDto:
      type: object
      required:
        - amount
        - mdasId
        - narration
      properties:
        amount:
          type: number
          format: double
        mdasId:
          type: integer
          description: Revenue Head ID
        narration:
          type: string
        recommendedRate:
          type: number
          format: double
    BulkBillRequest:
      type: object
      properties:
        engineCode:
          type: string
        esBillDtos:
          type: array
          items:
            allOf:
              - $ref: '#/components/schemas/CreateBillRequest'
              - type: object
                properties:
                  engineCode:
                    type: string
    Bill:
      type: object
      properties:
        billReference:
          type: string
        payStatus:
          type: string
        narration:
          type: string
        head:
          type: string
          nullable: true
          description: Revenue head identifier
        subhead:
          type: string
          nullable: true
          description: Revenue subhead identifier
        paidAt:
          type: string
          format: date-time
          nullable: true
          description: Timestamp when bill was paid
    BillItem:
      type: object
      properties:
        revenueHead:
          type: string
        amount:
          type: number
          format: double
        revenueCode:
          type: string
    CreateBillResponse:
      type: object
      properties:
        bill:
          $ref: '#/components/schemas/Bill'
        billItems:
          type: array
          items:
            $ref: '#/components/schemas/BillItem'
        failedBillItems:
          type: array
          items:
            $ref: '#/components/schemas/BillItem'
    AttachDataRequest:
      type: object
      required:
        - billReference
        - additionalData
      properties:
        billReference:
          type: string
        additionalData:
          type: object
          additionalProperties: true
    BulkAttachDataRequest:
      type: object
      required:
        - bills
      properties:
        bills:
          type: array
          items:
            type: object
            required:
              - billReference
              - additionalData
            properties:
              billReference:
                type: string
              additionalData:
                type: object
                additionalProperties: true
    InvoiceUrlResponse:
      type: object
      properties:
        invoiceUrl:
          type: string
          format: uri
...
    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: string
        statusCode:
          type: integer
    ErrorResponse:
      type: object
      properties:
        error:
          $ref: '#/components/schemas/Error'
    WebhookEvent:
      type: object
      required:
        - event
        - data
        - message
      properties:
        event:
          type: string
          description: Event type (e.g., 'charge.success', 'payment.success', 'payment.error')
          example: "charge.success"
        data:
          type: object
          description: Event data (must include invoiceNo)
          additionalProperties: true
          properties:
            invoiceNo:
              type: string
              description: Invoice number
              example: "INV123456"
          example:
            invoiceNo: "INV123456"
            billReference: "BILL123456"
            amount: 10000.00
            status: "paid"
        message:
          type: string
          description: Event message
          example: "Payment completed successfully"
...
  responses:
    BadRequest:
      description: Validation error or PayKaduna rejection
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Invalid API key or HMAC signature
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Bill or taxpayer not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalServerError:
      description: Server error or network failure
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'